+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and a race condition in otp_socket code
 - Fixed a bug where calling repeatedly `process_info` on a stopped process could cause an out of
 memory error
 - Fixed possible concurrency problems in ESP32 UART driver
+- Fixed concurrency and memory leak related to links and monitors
 
 ### Changed